+++ aat-gtk/src/main/kotlin/ch/bailu/aat_gtk/view/map/control/MainBar.kt
@@ -7,8 +7,11 @@ import ch.bailu.aat_gtk.controller.UiControllerInterface
 import ch.bailu.aat_gtk.search.SearchController
 import ch.bailu.aat_gtk.view.toplevel.navigation.NavigationView
 import ch.bailu.aat_gtk.view.toplevel.navigation.NavigationViewChanged
+import ch.bailu.aat_gtk.view.tracker.TrackerControllerView
+import ch.bailu.aat_lib.dispatcher.DispatcherInterface
 import ch.bailu.aat_lib.file.json.SearchModel
 import ch.bailu.aat_lib.map.edge.Position
+import ch.bailu.aat_lib.service.ServicesInterface
 import ch.bailu.aat_lib.util.extensions.ellipsize
 import ch.bailu.gtk.gio.Menu
 import ch.bailu.gtk.gtk.Application
@@ -22,22 +25,23 @@ import ch.bailu.gtk.gtk.Separator
 import ch.bailu.gtk.gtk.Widget
 import ch.bailu.gtk.lib.handler.action.ActionHandler
 
-class MainBar(private val app: Application, controller: UiControllerInterface) : Bar(Position.TOP),
+class MainBar(private val app: Application, uiController: UiControllerInterface, services: ServicesInterface, dispatcher: DispatcherInterface) : Bar(Position.TOP),
     NavigationViewChanged {
     private val searchModel = SearchModel()
     private val searchController = SearchController(searchModel)
+    private val trackerControllerView = TrackerControllerView(services, dispatcher, uiController)
     private val leftNavigationButton = Button().apply {
         addCssClass(Strings.mapControl)
         iconName = Icons.goPreviousSymbolic
         onClicked {
-            controller.hideMap()
+            uiController.hideMap()
         }
     }
 
     private val rightNavigationButton = Button().apply {
         setLabel("POI")
         onClicked {
-            controller.showPoi()
+            uiController.showPoi()
         }
     }
 
@@ -46,21 +50,25 @@ class MainBar(private val app: Application, controller: UiControllerInterface) :
 
     init {
         searchController.centerMap = { pos ->
-            controller.centerInMap(pos)
+            uiController.centerInMap(pos)
         }
 
-        add(Box(Orientation.HORIZONTAL, Layout.MARGIN).apply {
-            append(leftNavigationButton)
-            append(leftSeparator)
-            append(createNominatimWidget())
-            append(rightSeparator)
-            append(rightNavigationButton)
+        add(Box(Orientation.VERTICAL, Layout.MARGIN).apply {
+            append(Box(Orientation.HORIZONTAL, Layout.MARGIN).apply {
+                append(leftNavigationButton)
+                append(leftSeparator)
+                append(createNominatimWidget())
+                append(rightSeparator)
+                append(rightNavigationButton)
+            })
+            append(trackerControllerView.box)
         })
     }
 
     override fun onNavigationViewChanged(navigationView: NavigationView) {
         leftSeparator.visible = navigationView.leftCollapsed
         leftNavigationButton.visible = navigationView.leftCollapsed
+        trackerControllerView.box.visible = navigationView.leftCollapsed
         rightSeparator.visible = navigationView.rightCollapsed
         rightNavigationButton.visible = navigationView.rightCollapsed
     }